Enhanced files: src/s_serv.c
Only reports STATS requests to +y users when the STATS parameter is a
valid option.  A new integer is introduced into the m_stats() function
called 'valid_stats'.  If 'valid_stats' exists, it will report the
STATS usage.  In the future if it is decided to add support for multiple
STATS requests in one query, this will come in handy in measuring the
number of STATS requests.

-- David-R <rientjes@mail.whidbey.net>

fixed m_join() so it does not try to join to the channel if you
are already on it.

committed by db