admin.patch (904 bytes)
Enhanced files: tools/chkconf.c; src/s_conf.c

This patch removes the possibility of accepting a bad A:line.  Previously
an A:line would be accepted regardless of how many fields it contained.
We now make sure we have at least three fields to report and avoid these
type of occurances:

 Administrative info for shell.whidbey.net
 {null}
 {null}

It also adds a check to 'chkconf' to report this type of error ahead of
time.

-- David-R <rientjes@mail.whidbey.net>